Split out Router from the core Handlers code
