+++ src/HTTP.jl
@@ -601,6 +601,7 @@ include("download.jl")
 
 include("Handlers.jl")                 ;using .Handlers
 include("Servers.jl"); using .Servers; using .Servers: listen
+include("Routers.jl")                  ;using .Routers
 include("WebSockets.jl")               ;using .WebSockets
 
 import .ConnectionPool: Transaction, Connection
